build: build react packages with vite, node+browser packages with esbuild
