+++ package.json
@@ -19,8 +19,8 @@
     "prepare": "pnpm create-config",
     "build": "pnpm turbo run build --scope='@nhost/*' --no-deps --include-dependencies",
     "build:docs": "pnpm run build --filter=nhost-documentation",
-    "clean:all": "pnpm clean && rm -rf ./packages/*/node_modules node_modules",
-    "clean": "rm -rf ./packages/*/dist",
+    "clean:all": "pnpm clean && rm -rf ./{{packages,examples}/*,docs}/node_modules node_modules",
+    "clean": "rm -rf ./{{packages,examples}/*,docs}/{dist,.next,.turbo,coverage}",
     "ci": "pnpm turbo run build test --concurrency=4 --scope='@nhost/*' && pnpm run lint",
     "ci:version": "changeset version && pnpm install --frozen-lockfile false",
     "coverage": "pnpm run test --coverage",
@@ -50,6 +50,8 @@
     "@typescript-eslint/eslint-plugin": "^5.10.2",
     "@typescript-eslint/parser": "^5.10.2",
     "@vitejs/plugin-react": "^1.1.4",
+    "esbuild": "^0.14.23",
+    "esbuild-node-externals": "^1.4.1",
     "eslint": "^8.8.0",
     "eslint-config-react-app": "^7.0.0",
     "eslint-plugin-flowtype": "^8.0.3",
