+++ packages/nhost-js/src/clients/functions.ts
@@ -1,8 +1,6 @@
-import type { AxiosInstance, AxiosRequestConfig } from 'axios'
+import axios, { AxiosInstance, AxiosRequestConfig } from 'axios'
 
 import { FunctionCallResponse } from '../types'
-
-const axios = require('axios')
 export interface NhostFunctionsConstructorParams {
   url: string
 }
