+++ packages/nhost-js/src/clients/graphql.ts
@@ -1,8 +1,6 @@
-import type { AxiosInstance, AxiosRequestConfig } from 'axios'
+import axios, { AxiosInstance, AxiosRequestConfig } from 'axios'
 
 import { GraphqlRequestResponse, GraphqlResponse } from '../types'
-
-const axios = require('axios')
 
 export interface NhostGraphqlConstructorParams {
   url: string
